rag-project-1/requirements.txt
chromadb
pypdf
google-genai
python-dotenv
nltk
